# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Control de Presupuesto",
    "description": "Esta API proporciona acceso a las funcionalidades de Control de Presupuesto ",
    "version": "1.0.0",
    "contact": {
      "name": "Esmir Roque",
      "email": "castellano17@gmail.com",
      "url": "https://www.linkedin.com/in/esmir-roque/"
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
...
    "/api/v1/users/": {
      "post": {
        "tags": ["Usuarios"],
        "summary": "Crea un nuevo usuario",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "description": "El nombre del usuario"
                  },
                  "lastName": {
                    "type": "string",
                    "description": "El apellido del usuario"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "el correo del usuario"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "la contraseña del usuario"
                  },
                  "birthday": {
                    "type": "string",
                    "format": "date",
                    "description": "La fecha del cumpleaños del usuario"
                  },
                  "phone": {
                    "type": "string",
                    "description": "El telefono del usuario"
                  },
                  "country": {
                    "type": "string",
                    "description": "El pais del usuario"
                  },
                  "gender": {
                    "type": "string",
                    "description": "El género del usuario"
                  }
                },

                "required": ["firstName", "lastName", "email", "password"]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Al hacer esta petición agregamos un usuario nuevo"
          },
          "400": {
            "description": "Bad Request: La petición es inválida o faltan parámetros"
          }
        }
      }
    },
...
    "/api/v1/category": {
      "get": {
        "tags": ["Categorías"],
        "summary": "Muestra las categorías",
        "security": [
          {
            "BearerAuth": [],
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Al hacer esta petición debería retornar las categorías"
          },
          "401": {
            "description": "Error: Unauthorized "
          }
        }
      }
    },
...
    "/api/v1/expenses": {
      "get": {
        "tags": ["Gastos"],
        "summary": "Muestra los gastos del usuario logueado",
        "security": [
          {
            "BearerAuth": [],
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Al hacer esta petición debería retornar los gastos del usuario logueado"
          },
          "401": {
            "description": "Error: Unauthorized "
          }
        }
      }
    },
...
  }
}